requirements.txt
aiofiles>=0.6.0
APScheduler>=3.7.0
Pillow>=8.1.1
nonebot2>=2.0.0b1
nonebot-adapter-onebot>=2.0.0b1
nonebug>=0.2.0
nonebot-plugin-gocqhttp>=0.3.2
psutil>=5.7.2
pathlib>=1.0.1
pytz>=2020.1
pyyaml>=5.4
scikit-image>=0.18.3
tensorflow>=2.6.0